Module substrate_network::config

Configuration of the networking layer.

The [Params] struct is the struct that must be passed in order to initialize the networking. See the documentation of [Params].

Modules

identity
A node's network identity keys.

Macros

build_multiaddr
Easy way for a user to create a Multiaddr.

Structs

DummyFinalityProofRequestBuilder
Implementation of FinalityProofRequestBuilder that builds a dummy empty request.

ExtTransport
Implementation of Transport whose implementation is handled by some FFI.

NetworkConfiguration
Network service configuration.

Params
Network initialization parameters.

ProtocolConfig
Configuration for the Substrate-specific part of the networking layer.

ProtocolId
Name of a protocol, transmitted on the wire. Should be unique for each chain.

Roles
Bitmask of the roles that a node fulfills.

Enums

NodeKeyConfig
The configuration of a node's secret key, describing the type of key and how it is obtained. A node's identity keypair is the result of the evaluation of the node key configuration.

NonReservedPeerMode
The policy for connections to non-reserved peers.

ParseErr
Error that can be generated by parse_str_addr.

PublicKey
The public key of a node's identity keypair.

Secret
The configuration options for obtaining a secret key K.

TransportConfig
Configuration for the transport layer.

Traits

FinalityProofRequestBuilder
Finality proof request builder.

Functions

parse_addr
Splits a Multiaddress into a Multiaddress and PeerId.

parse_str_addr
Parses a string address and splits it into Multiaddress and PeerId, if valid.

Type Definitions

BoxFinalityProofRequestBuilder
Shared finality proof request builder struct used by the queue.

Ed25519Secret
The options for obtaining a Ed25519 secret key.

Secp256k1Secret
The options for obtaining a Secp256k1 secret key.