Struct substrate_network::config::NetworkConfiguration

pub struct NetworkConfiguration {
    pub config_path: Option<String>,
    pub net_config_path: Option<String>,
    pub listen_addresses: Vec<Multiaddr>,
    pub public_addresses: Vec<Multiaddr>,
    pub boot_nodes: Vec<String>,
    pub node_key: NodeKeyConfig,
    pub in_peers: u32,
    pub out_peers: u32,
    pub reserved_nodes: Vec<String>,
    pub non_reserved_mode: NonReservedPeerMode,
    pub client_version: String,
    pub node_name: String,
    pub enable_mdns: bool,
    pub wasm_external_transport: Option<ExtTransport>,
}

Network service configuration.

Fields
config_path: Option<String>

Directory path to store general network configuration. None means nothing will be saved.

net_config_path: Option<String>

Directory path to store network-specific configuration. None means nothing will be saved.

listen_addresses: Vec<Multiaddr>

Multiaddresses to listen for incoming connections.

public_addresses: Vec<Multiaddr>

Multiaddresses to advertise. Detected automatically if empty.

boot_nodes: Vec<String>

List of initial node addresses

node_key: NodeKeyConfig

The node key configuration, which determines the node's network identity keypair.

in_peers: u32

Maximum allowed number of incoming connections.

out_peers: u32

Number of outgoing connections we're trying to maintain.

reserved_nodes: Vec<String>

List of reserved node addresses.

non_reserved_mode: NonReservedPeerMode

The non-reserved peer mode.

client_version: String

Client identifier. Sent over the wire for debugging purposes.

node_name: String

Name of the node. Sent over the wire for debugging purposes.

enable_mdns: bool

If true, the network will use mDNS to discover other libp2p nodes on the local network and connect to them if they support the same chain.

wasm_external_transport: Option<ExtTransport>

Optional external implementation of a libp2p transport. Used in WASM contexts where we need some binding between the networking provided by the operating system or environment and libp2p.

This parameter exists whatever the target platform is, but it is expected to be set to Some only when compiling for WASM.

Methods

impl NetworkConfiguration

pub fn new() -> Self

Create a new instance of default settings.

pub fn new_local() -> NetworkConfiguration

Create new default configuration for localhost-only connection with random port (useful for testing)