Macro substrate_network::construct_simple_protocol

macro_rules! construct_simple_protocol {
    (
		$( #[ $attr:meta ] )*
		pub struct $protocol:ident where
			Block = $block:ident
			$( , Status = $status_protocol_name:ident )*
		{
			$( $sub_protocol_name:ident : $sub_protocol:ident $( <$protocol_block:ty> )*, )*
		}
	) => { ... };
}

Construct a simple protocol that is composed of several sub protocols. Each "sub protocol" needs to implement Specialization and needs to provide a new() function. For more fine grained implementations, this macro is not usable.

Example

construct_simple_protocol! {
    pub struct MyProtocol where Block = MyBlock {
        consensus_gossip: ConsensusGossip<MyBlock>,
        other_protocol: MyCoolStuff,
    }
}

You can also provide an optional parameter after where Block = MyBlock, so it looks like where Block = MyBlock, Status = consensus_gossip. This will instruct the implementation to use the status() function from the ConsensusGossip protocol. By default, status() returns an empty vector.