Struct substrate_network::multiaddr::multihash::Multihash

pub struct Multihash { /* fields omitted */ }

Represents a valid multihash.

Methods

impl Multihash

pub fn from_bytes(bytes: Vec<u8>) -> Result<Multihash, DecodeOwnedError>

Verifies whether bytes contains a valid multihash, and if so returns a Multihash.

pub fn random(hash: Hash) -> Multihash

Generates a random Multihash from a cryptographically secure PRNG.

pub fn into_bytes(self) -> Vec<u8>

Returns the bytes representation of the multihash.

pub fn to_vec(&self) -> Vec<u8>

Returns the bytes representation of the multihash.

pub fn as_bytes(&self) -> &[u8]

Returns the bytes representation of this multihash.

pub fn as_ref(&self) -> MultihashRef

Builds a MultihashRef corresponding to this Multihash.

pub fn algorithm(&self) -> Hash

Returns which hashing algorithm is used in this multihash.

pub fn digest(&self) -> &[u8]

Returns the hashed data.