Trait substrate_network::specialization::NetworkSpecialization

pub trait NetworkSpecialization<B: BlockT>: Send + Sync + 'static {
    fn status(&self) -> Vec<u8>;

    fn on_connect(
        &mut self, 
        ctx: &mut dyn Context<B>, 
        who: PeerId, 
        status: Status<B>
    );

    fn on_disconnect(&mut self, ctx: &mut dyn Context<B>, who: PeerId);

    fn on_message(
        &mut self, 
        ctx: &mut dyn Context<B>, 
        who: PeerId, 
        message: Vec<u8>
    );

    fn on_event(&mut self, event: Event);

    fn on_abort(&mut self) { ... }

    fn maintain_peers(&mut self, _ctx: &mut dyn Context<B>) { ... }

    fn on_block_imported(
        &mut self, 
        _ctx: &mut dyn Context<B>, 
        _hash: B::Hash, 
        _header: &B::Header
    ) { ... }
}

A specialization of the substrate network protocol. Handles events and sends messages.

Required methods

fn status(&self) -> Vec<u8>

Get the current specialization-status.

fn on_connect(
&mut self,
ctx: &mut dyn Context<B>,
who: PeerId,
status: Status<B>
)

Called when a peer successfully handshakes.

fn on_disconnect(&mut self, ctx: &mut dyn Context<B>, who: PeerId)

Called when a peer is disconnected. If the peer ID is unknown, it should be ignored.

fn on_message(
&mut self,
ctx: &mut dyn Context<B>,
who: PeerId,
message: Vec<u8>
)

Called when a network-specific message arrives.

fn on_event(&mut self, event: Event)

Called when a network-specific event arrives.

Loading content...
Provided methods

fn on_abort(&mut self)

Deprecated:
This method is never called; aborting corresponds to dropping the object

Called on abort.

fn maintain_peers(&mut self, _ctx: &mut dyn Context<B>)

Called periodically to maintain peers and handle timeouts.

fn on_block_imported(
&mut self,
_ctx: &mut dyn Context<B>,
_hash: B::Hash,
_header: &B::Header
)

Called when a block is imported at the head of the chain (not during major sync). Not guaranteed to be called for every block, but will be most of the after major sync.

Loading content...