Struct substrate_network::NetworkService

pub struct NetworkService<B: BlockT + 'static, S: NetworkSpecialization<B>, H: ExHashT> { /* fields omitted */ }

Substrate network service. Handles network IO and manages connectivity.

Methods

impl<B: BlockT + 'static, S: NetworkSpecialization<B>, H: ExHashT> NetworkService<B, S, H>

pub fn local_peer_id(&self) -> PeerId

Returns the network identity of the node.

pub fn trigger_repropagate(&self)

You must call this when new transactons are imported by the transaction pool.

The latest transactions will be fetched from the TransactionPool that was passed at initialization as part of the configuration.

pub fn announce_block(&self, hash: B::Hash)

Make sure an important block is propagated to peers.

In chain-based consensus, we often need to make sure non-best forks are at least temporarily synced. This function forces such an announcement.

pub fn gossip_consensus_message(
&self,
topic: B::Hash,
engine_id: ConsensusEngineId,
message: Vec<u8>,
recipient: GossipMessageRecipient
)

Send a consensus message through the gossip

pub fn report_peer(&self, who: PeerId, cost_benefit: i32)

Report a given peer as either beneficial (+) or costly (-) according to the given scalar.

pub fn request_justification(&self, hash: &B::Hash, number: NumberFor<B>)

Request a justification for the given block from the network.

On success, the justification will be passed to the import queue that was part at initialization as part of the configuration.

pub fn with_spec<F>(&self, f: F) where
F: FnOnce(&mut S, &mut dyn Context<B>) + Send + 'static,

Execute a closure with the chain-specific network specialization.

pub fn with_gossip<F>(&self, f: F) where
F: FnOnce(&mut ConsensusGossip<B>, &mut dyn Context<B>) + Send + 'static,

Execute a closure with the consensus gossip.

pub fn is_major_syncing(&self) -> bool

Are we in the process of downloading the chain?

pub fn get_value(&self, key: &Multihash)

Start getting a value from the DHT.

This will generate either a ValueFound or a ValueNotFound event and pass it to on_event on the network specialization.

pub fn put_value(&self, key: Multihash, value: Vec<u8>)

Start putting a value in the DHT.

This will generate either a ValuePut or a ValuePutFailed event and pass it to on_event on the network specialization.

pub fn accept_unreserved_peers(&self)

Connect to unreserved peers and allow unreserved peers to connect.

pub fn deny_unreserved_peers(&self)

Disconnect from unreserved peers and deny new unreserved peers to connect.

pub fn remove_reserved_peer(&self, peer: PeerId)

Removes a PeerId from the list of reserved peers.

pub fn add_reserved_peer(&self, peer: String) -> Result<(), String>

Adds a PeerId and its address as reserved.

pub fn num_connected(&self) -> usize

Returns the number of peers we're connected to.

pub fn external_addresses(&self) -> Vec<Multiaddr>

Returns the local external addresses.