Struct substrate_network::NetworkWorker

#[must_use = "The NetworkWorker must be polled in order for the network to work"]
pub struct NetworkWorker<B: BlockT + 'static, S: NetworkSpecialization<B>, H: ExHashT> { /* fields omitted */ }

Main network worker. Must be polled in order for the network to advance.

You are encouraged to poll this in a separate background thread or task.

Methods

impl<B: BlockT + 'static, S: NetworkSpecialization<B>, H: ExHashT> NetworkWorker<B, S, H>

pub fn new(params: Params<B, S, H>) -> Result<NetworkWorker<B, S, H>, Error>

Creates the network service.

Returns a NetworkWorker that implements Future and must be regularly polled in order for the network processing to advance. From it, you can extract a NetworkService using worker.service(). The NetworkService can be shared through the codebase.

pub fn average_download_per_sec(&self) -> u64

Returns the downloaded bytes per second averaged over the past few seconds.

pub fn average_upload_per_sec(&self) -> u64

Returns the uploaded bytes per second averaged over the past few seconds.

pub fn num_connected_peers(&self) -> usize

Returns the number of peers we're connected to.

pub fn num_active_peers(&self) -> usize

Returns the number of peers we're connected to and that are being queried.

pub fn sync_state(&self) -> SyncState

Current global sync state.

pub fn best_seen_block(&self) -> Option<NumberFor<B>>

Target sync block number.

pub fn num_sync_peers(&self) -> u32

Number of peers participating in syncing.

pub fn num_queued_blocks(&self) -> u32

Number of blocks in the import queue.

pub fn add_known_address(&mut self, peer_id: PeerId, addr: Multiaddr)

Adds an address for a node.

pub fn service(&self) -> &Arc<NetworkService<B, S, H>>

Return a NetworkService that can be shared through the code base and can be used to manipulate the worker.

pub fn on_block_imported(
&mut self,
hash: B::Hash,
header: B::Header,
data: Vec<u8>,
is_best: bool
)

You must call this when a new block is imported by the client.

pub fn on_block_finalized(&mut self, hash: B::Hash, header: B::Header)

You must call this when a new block is finalized by the client.

pub fn network_state(&mut self) -> NetworkState

Get network state.

Note: Use this only for debugging. This API is unstable. There are warnings literaly everywhere about this. Please don't use this function to retrieve actual information.

pub fn peers_debug_info(&mut self) -> Vec<(PeerId, PeerInfo<B>)>

Get currently connected peers.