Struct substrate_network::Service

pub struct Service<B: BlockT + 'static, S: NetworkSpecialization<B>> { /* fields omitted */ }

Substrate network service. Handles network IO and manages connectivity.

Methods

impl<B: BlockT + 'static, S: NetworkSpecialization<B>> Service<B, S>

pub fn new<H: ExHashT>(
params: Params<B, S, H>,
protocol_id: ProtocolId,
import_queue: Box<dyn ImportQueue<B>>
) -> Result<(Arc<Service<B, S>>, NetworkChan<B>), Error>

Creates and register protocol with the network service

pub fn average_download_per_sec(&self) -> u64

Returns the downloaded bytes per second averaged over the past few seconds.

pub fn average_upload_per_sec(&self) -> u64

Returns the uploaded bytes per second averaged over the past few seconds.

pub fn local_peer_id(&self) -> PeerId

Returns the network identity of the node.

pub fn on_block_imported(&self, hash: B::Hash, header: B::Header)

Called when a new block is imported by the client.

pub fn on_block_finalized(&self, hash: B::Hash, header: B::Header)

Called when a new block is finalized by the client.

pub fn trigger_repropagate(&self)

Called when new transactons are imported by the client.

pub fn announce_block(&self, hash: B::Hash)

Make sure an important block is propagated to peers.

In chain-based consensus, we often need to make sure non-best forks are at least temporarily synced.

pub fn gossip_consensus_message(
&self,
topic: B::Hash,
engine_id: ConsensusEngineId,
message: Vec<u8>,
force: bool
)

Send a consensus message through the gossip

pub fn with_spec<F>(&self, f: F) where
F: FnOnce(&mut S, &mut dyn Context<B>) + Send + 'static,

Execute a closure with the chain-specific network specialization.

pub fn with_gossip<F>(&self, f: F) where
F: FnOnce(&mut ConsensusGossip<B>, &mut dyn Context<B>) + Send + 'static,

Execute a closure with the consensus gossip.