Module substrate_network::test

Structs

DummySpecialization
The test specialization.

EmptyTransactionPool

ForceFinalized

JustificationTestNet

NetworkLink
A link implementation that connects to the network.

NoopLink
A link implementation that does nothing.

PassThroughVerifier
A Verifier that accepts all blocks and passes them on with the configured finality to be imported.

Peer

TestLink
A Link that can wait for a block to have been imported.

TestNet

Transfer
Calls in transactions.

Enums

Extrinsic
Extrinsic for test-runtime.

FromNetworkMsg

PeersClient

Traits

SpecializationFactory

TestNetFactory

Type Definitions

Block
A test block.

Hash
A simple hash type for all our hashing.

PeersFullClient

PeersLightClient

TestClient
A test client with default backend.