Struct substrate_network::test::Peer

pub struct Peer<D, S: NetworkSpecialization<Block> + Clone> {
    pub is_offline: Arc<AtomicBool>,
    pub is_major_syncing: Arc<AtomicBool>,
    pub peers: Arc<RwLock<HashMap<PeerId, ConnectedPeer<Block>>>>,
    pub peer_id: PeerId,
    pub import_queue: Box<BasicQueue<Block>>,
    pub data: D,
    // some fields omitted
}

Fields
is_offline: Arc<AtomicBool>is_major_syncing: Arc<AtomicBool>peers: Arc<RwLock<HashMap<PeerId, ConnectedPeer<Block>>>>peer_id: PeerIdimport_queue: Box<BasicQueue<Block>>data: D
Methods

impl<D, S: NetworkSpecialization<Block> + Clone> Peer<D, S>

pub fn on_block_imported(
&self,
hash: <Block as BlockT>::Hash,
header: &<Block as BlockT>::Header
)

pub fn send_finality_notifications(&self)

Send block finalization notifications.

pub fn gossip_message(
&self,
topic: <Block as BlockT>::Hash,
engine_id: ConsensusEngineId,
data: Vec<u8>,
force: bool
)

Push a message into the gossip network and relay to peers. TestNet::sync_step needs to be called to ensure it's propagated.

pub fn consensus_gossip_collect_garbage_for_topic(
&self,
_topic: <Block as BlockT>::Hash
)

pub fn consensus_gossip_messages_for(
&self,
engine_id: ConsensusEngineId,
topic: <Block as BlockT>::Hash
) -> UnboundedReceiver<Vec<u8>>

access the underlying consensus gossip handler

pub fn with_gossip<F>(&self, f: F) where
F: FnOnce(&mut ConsensusGossip<Block>, &mut dyn Context<Block>) + Send + 'static,

Execute a closure with the consensus gossip.

pub fn announce_block(&self, block: Hash)

Announce a block to peers.

pub fn generate_blocks<F>(
&self,
count: usize,
origin: BlockOrigin,
edit_block: F
) -> H256 where
F: FnMut(BlockBuilder<Block, PeersClient>) -> Block,

Add blocks to the peer -- edit the block before adding

pub fn generate_blocks_at<F>(
&self,
at: BlockId<Block>,
count: usize,
origin: BlockOrigin,
edit_block: F
) -> H256 where
F: FnMut(BlockBuilder<Block, PeersClient>) -> Block,

Add blocks to the peer -- edit the block before adding. The chain will start at the given block iD.

pub fn push_blocks(&self, count: usize, with_tx: bool) -> H256

Push blocks to the peer (simplified: with or without a TX)

pub fn push_blocks_at(
&self,
at: BlockId<Block>,
count: usize,
with_tx: bool
) -> H256

Push blocks to the peer (simplified: with or without a TX) starting from given hash.

pub fn push_authorities_change_block(
&self,
new_authorities: Vec<AuthorityId>
) -> H256

pub fn client(&self) -> &Arc<PeersClient>

Get a reference to the client.