Trait substrate_network::test::TestClientBuilderExt

pub trait TestClientBuilderExt<B> {
    fn set_support_changes_trie(self, support_changes_trie: bool) -> Self;

    fn set_heap_pages(self, heap_pages: u64) -> Self;

    fn add_extra_child_storage<SK, K, V>(
        self, 
        storage_key: SK, 
        key: K, 
        value: V
    ) -> Self
    where
        K: Into<Vec<u8>>,
        SK: Into<Vec<u8>>,
        V: Into<Vec<u8>>;

    fn add_extra_storage<K, V>(self, key: K, value: V) -> Self
    where
        K: Into<Vec<u8>>,
        V: Into<Vec<u8>>;

    fn build_with_longest_chain(
        self
    ) -> (Client<B, LocalCallExecutor<B, NativeExecutor<LocalExecutor>>, Block<Header<u64, BlakeTwo256>, Extrinsic>, RuntimeApi>, LongestChain<B, Block<Header<u64, BlakeTwo256>, Extrinsic>>);

    fn build(
        self
    ) -> Client<B, LocalCallExecutor<B, NativeExecutor<LocalExecutor>>, Block<Header<u64, BlakeTwo256>, Extrinsic>, RuntimeApi> { ... }
}

A test-runtime extensions to TestClientBuilder.

Required methods

fn set_support_changes_trie(self, support_changes_trie: bool) -> Self

Enable or disable support for changes trie in genesis.

fn set_heap_pages(self, heap_pages: u64) -> Self

Override the default value for Wasm heap pages.

fn add_extra_child_storage<SK, K, V>(
self,
storage_key: SK,
key: K,
value: V
) -> Self where
K: Into<Vec<u8>>,
SK: Into<Vec<u8>>,
V: Into<Vec<u8>>,

Add an extra value into the genesis storage.

Panics

Panics if the key is empty.

fn add_extra_storage<K, V>(self, key: K, value: V) -> Self where
K: Into<Vec<u8>>,
V: Into<Vec<u8>>,

Add an extra child value into the genesis storage.

Panics

Panics if the key is empty.

fn build_with_longest_chain(
self
) -> (Client<B, LocalCallExecutor<B, NativeExecutor<LocalExecutor>>, Block<Header<u64, BlakeTwo256>, Extrinsic>, RuntimeApi>, LongestChain<B, Block<Header<u64, BlakeTwo256>, Extrinsic>>)

Build the test client and longest chain selector.

Loading content...
Provided methods

fn build(
self
) -> Client<B, LocalCallExecutor<B, NativeExecutor<LocalExecutor>>, Block<Header<u64, BlakeTwo256>, Extrinsic>, RuntimeApi>

Build the test client.

Loading content...