Trait substrate_network::test::TestNetFactory

pub trait TestNetFactory: Sized {
    type Specialization: NetworkSpecialization<Block> + SpecializationFactory;
    type Verifier: 'static + Verifier<Block>;
    type PeerData: Default;
    fn from_config(config: &ProtocolConfig) -> Self;

    fn make_verifier(
        &self, 
        client: PeersClient, 
        config: &ProtocolConfig
    ) -> Arc<Self::Verifier>;

    fn peer(&self, i: usize) -> &Peer<Self::PeerData, Self::Specialization>;

    fn peers(&self) -> &Vec<Arc<Peer<Self::PeerData, Self::Specialization>>>;

    fn mut_peers<F: FnOnce(&mut Vec<Arc<Peer<Self::PeerData, Self::Specialization>>>)>(
        &mut self, 
        closure: F
    );

    fn started(&self) -> bool;

    fn set_started(&mut self, now: bool);

    fn make_block_import(
        &self, 
        client: PeersClient
    ) -> (SharedBlockImport<Block>, Option<SharedJustificationImport<Block>>, Option<SharedFinalityProofImport<Block>>, Option<SharedFinalityProofRequestBuilder<Block>>, Self::PeerData) { ... }

    fn make_finality_proof_provider(
        &self, 
        _client: PeersClient
    ) -> Option<Arc<dyn FinalityProofProvider<Block>>> { ... }

    fn default_config() -> ProtocolConfig { ... }

    fn uses_tokio(&self) -> bool { ... }

    fn new(n: usize) -> Self { ... }

    fn add_peer(
        &mut self, 
        protocol_status: Arc<RwLock<(bool, bool, usize)>>, 
        import_queue: Arc<Mutex<Box<BasicQueue<Block>>>>, 
        tx_pool: EmptyTransactionPool, 
        finality_proof_provider: Option<Arc<dyn FinalityProofProvider<Block>>>, 
        protocol: Protocol<Block, Self::Specialization, Hash>, 
        protocol_sender: UnboundedSender<ProtocolMsg<Block, Self::Specialization>>, 
        network_to_protocol_sender: UnboundedSender<FromNetworkMsg<Block>>, 
        network_sender: UnboundedSender<NetworkMsg<Block>>, 
        network_to_protocol_rx: UnboundedReceiver<FromNetworkMsg<Block>>, 
        protocol_rx: UnboundedReceiver<ProtocolMsg<Block, Self::Specialization>>, 
        peer: Arc<Peer<Self::PeerData, Self::Specialization>>
    ) { ... }

    fn add_full_peer(&mut self, config: &ProtocolConfig) { ... }

    fn add_light_peer(&mut self, config: &ProtocolConfig) { ... }

    fn start(&mut self) { ... }

    fn route_single(
        &mut self, 
        disconnect: bool, 
        disconnected: Option<HashSet<usize>>, 
        message_filter: &dyn Fn(&NetworkMsg<Block>) -> bool
    ) -> bool { ... }

    fn send_import_notifications(&mut self) { ... }

    fn send_finality_notifications(&mut self) { ... }

    fn sync_with(
        &mut self, 
        disconnect: bool, 
        disconnected: Option<HashSet<usize>>
    ) { ... }

    fn sync_step(&mut self) { ... }

    fn tick_peer(&mut self, i: usize) { ... }

    fn sync(&mut self) { ... }

    fn sync_without_disconnects(&mut self) { ... }

    fn done(&self) -> bool { ... }
}

Associated Types

type Specialization: NetworkSpecialization<Block> + SpecializationFactory

type Verifier: 'static + Verifier<Block>

type PeerData: Default

Loading content...
Required methods

fn from_config(config: &ProtocolConfig) -> Self

These two need to be implemented!

fn make_verifier(
&self,
client: PeersClient,
config: &ProtocolConfig
) -> Arc<Self::Verifier>

fn peer(&self, i: usize) -> &Peer<Self::PeerData, Self::Specialization>

Get reference to peer.

fn peers(&self) -> &Vec<Arc<Peer<Self::PeerData, Self::Specialization>>>

fn mut_peers<F: FnOnce(&mut Vec<Arc<Peer<Self::PeerData, Self::Specialization>>>)>(
&mut self,
closure: F
)

fn started(&self) -> bool

fn set_started(&mut self, now: bool)

Loading content...
Provided methods

fn make_block_import(
&self,
client: PeersClient
) -> (SharedBlockImport<Block>, Option<SharedJustificationImport<Block>>, Option<SharedFinalityProofImport<Block>>, Option<SharedFinalityProofRequestBuilder<Block>>, Self::PeerData)

Get custom block import handle for fresh client, along with peer data.

fn make_finality_proof_provider(
&self,
_client: PeersClient
) -> Option<Arc<dyn FinalityProofProvider<Block>>>

Get finality proof provider (if supported).

fn default_config() -> ProtocolConfig

fn uses_tokio(&self) -> bool

Must return true if the testnet is going to be used from within a tokio context.

fn new(n: usize) -> Self

Create new test network with this many peers.

fn add_peer(
&mut self,
protocol_status: Arc<RwLock<(bool, bool, usize)>>,
import_queue: Arc<Mutex<Box<BasicQueue<Block>>>>,
tx_pool: EmptyTransactionPool,
finality_proof_provider: Option<Arc<dyn FinalityProofProvider<Block>>>,
protocol: Protocol<Block, Self::Specialization, Hash>,
protocol_sender: UnboundedSender<ProtocolMsg<Block, Self::Specialization>>,
network_to_protocol_sender: UnboundedSender<FromNetworkMsg<Block>>,
network_sender: UnboundedSender<NetworkMsg<Block>>,
network_to_protocol_rx: UnboundedReceiver<FromNetworkMsg<Block>>,
protocol_rx: UnboundedReceiver<ProtocolMsg<Block, Self::Specialization>>,
peer: Arc<Peer<Self::PeerData, Self::Specialization>>
)

Add created peer.

fn add_full_peer(&mut self, config: &ProtocolConfig)

Add a full peer.

fn add_light_peer(&mut self, config: &ProtocolConfig)

Add a light peer.

fn start(&mut self)

Start network.

fn route_single(
&mut self,
disconnect: bool,
disconnected: Option<HashSet<usize>>,
message_filter: &dyn Fn(&NetworkMsg<Block>) -> bool
) -> bool

Do single round of message routing: single message from every peer is routed.

fn send_import_notifications(&mut self)

Send block import notifications for all peers.

fn send_finality_notifications(&mut self)

Send block finalization notifications for all peers.

fn sync_with(&mut self, disconnect: bool, disconnected: Option<HashSet<usize>>)

Perform synchronization until complete, if provided the given nodes set are excluded from sync.

fn sync_step(&mut self)

Deliver at most 1 pending message from every peer.

fn tick_peer(&mut self, i: usize)

Maintain sync for a peer.

fn sync(&mut self)

Deliver pending messages until there are no more.

fn sync_without_disconnects(&mut self)

Deliver pending messages until there are no more. Do not disconnect nodes.

fn done(&self) -> bool

Whether all peers have no pending outgoing messages.

Loading content...