Module substrate_network::config

Configuration for the networking layer of Substrate.

Structs

NetworkConfiguration
Network service configuration.

Params
Service initialization parameters.

ProtocolConfig
Configuration for the Substrate-specific part of the networking layer.

Roles
Bitmask of the roles that a node fulfills.

Enums

NodeKeyConfig
The configuration of a node's secret key, describing the type of key and how it is obtained. A node's identity keypair is the result of the evaluation of the node key configuration.

NonReservedPeerMode
The policy for connections to non-reserved peers.

Secret
The configuration options for obtaining a secret key K.

Type Definitions

Ed25519Secret
The options for obtaining a Ed25519 secret key.

Secp256k1Secret
The options for obtaining a Secp256k1 secret key.