Struct substrate_network::consensus_gossip::ConsensusGossip

pub struct ConsensusGossip<B: BlockT> { /* fields omitted */ }

Consensus network protocol handler. Manages statements and candidate requests.

Methods

impl<B: BlockT> ConsensusGossip<B>

pub fn new() -> Self

Create a new instance.

pub fn abort(&mut self)

Closes all notification streams.

pub fn register_validator(
&mut self,
engine_id: ConsensusEngineId,
validator: Arc<dyn Validator<B::Hash>>
)

Register message validator for a message type.

pub fn new_peer(
&mut self,
protocol: &mut dyn Context<B>,
who: PeerId,
roles: Roles
)

Handle new connected peer.

pub fn peer_disconnected(&mut self, _protocol: &mut dyn Context<B>, who: PeerId)

Call when a peer has been disconnected to stop tracking gossip status.

pub fn collect_garbage(&mut self)

Prune old or no longer relevant consensus messages. Provide a predicate for pruning, which returns false when the items with a given topic should be pruned.

pub fn messages_for(
&mut self,
engine_id: ConsensusEngineId,
topic: B::Hash
) -> UnboundedReceiver<Vec<u8>>

Get data of valid, incoming messages for a topic (but might have expired meanwhile)

pub fn on_incoming(
&mut self,
protocol: &mut dyn Context<B>,
who: PeerId,
message: ConsensusMessage
) -> Option<(B::Hash, ConsensusMessage)>

Handle an incoming ConsensusMessage for topic by who via protocol. Discard message if topic already known, the message is old, its source peers isn't a registered peer or the connection to them is broken. Return Some(topic, message) if it was added to the internal queue, None in all other cases.

pub fn multicast(
&mut self,
protocol: &mut dyn Context<B>,
topic: B::Hash,
message: ConsensusMessage,
force: bool
)

Multicast a message to all peers.