Module substrate_network::message

Network packet message types. These get serialized and put into the lower level protocol payload.

Modules

generic
Generic types.

Structs

BlockAnnounce
Announce a new complete relay chain block on the network.

BlockAttributes
Node roles bitmask.

FinalityProofRequest
Finality proof request.

FinalityProofResponse
Finality proof response.

RemoteCallRequest
Remote call request.

RemoteCallResponse
Remote call response.

RemoteChangesRequest
Remote changes request.

RemoteChangesResponse
Remote changes response.

RemoteHeaderRequest
Remote header request.

RemoteHeaderResponse
Remote header response.

RemoteReadChildRequest
Remote storage read child request.

RemoteReadRequest
Remote storage read request.

RemoteReadResponse
Remote read response.

Enums

Direction
Block enumeration direction.

FromBlock
Identifies starting point of a block sequence.

Type Definitions

BlockData
Type alias for using the BlockData type using block type parameters.

BlockRequest
Type alias for using the block request type using block type parameters.

BlockResponse
Type alias for using the BlockResponse type using block type parameters.

Message
Type alias for using the message type using block type parameters.

RequestId
A unique ID of a request.

Status
Type alias for using the status type using block type parameters.

Transactions
A set of transactions.