Struct substrate_network::NetworkService

pub struct NetworkService<B: BlockT + 'static, S: NetworkSpecialization<B>> { /* fields omitted */ }

Substrate network service. Handles network IO and manages connectivity.

Methods

impl<B: BlockT + 'static, S: NetworkSpecialization<B>> NetworkService<B, S>

pub fn average_download_per_sec(&self) -> u64

Returns the downloaded bytes per second averaged over the past few seconds.

pub fn average_upload_per_sec(&self) -> u64

Returns the uploaded bytes per second averaged over the past few seconds.

pub fn local_peer_id(&self) -> PeerId

Returns the network identity of the node.

pub fn on_block_imported(&self, hash: B::Hash, header: B::Header)

Called when a new block is imported by the client.

pub fn on_block_finalized(&self, hash: B::Hash, header: B::Header)

Called when a new block is finalized by the client.

pub fn trigger_repropagate(&self)

Called when new transactons are imported by the client.

pub fn announce_block(&self, hash: B::Hash)

Make sure an important block is propagated to peers.

In chain-based consensus, we often need to make sure non-best forks are at least temporarily synced.

pub fn gossip_consensus_message(
&self,
topic: B::Hash,
engine_id: ConsensusEngineId,
message: Vec<u8>,
recipient: GossipMessageRecipient
)

Send a consensus message through the gossip

pub fn report_peer(&self, who: PeerId, cost_benefit: i32)

Report a given peer as either beneficial (+) or costly (-) according to the given scalar.

pub fn disconnect_peer(&self, who: PeerId)

Send a message to the given peer. Has no effect if we're not connected to this peer.

This method is extremely poor in terms of API and should be eventually removed.

pub fn send_request(&self, who: PeerId, message: Message<B>)

Send a message to the given peer. Has no effect if we're not connected to this peer.

This method is extremely poor in terms of API and should be eventually removed.

pub fn with_spec<F>(&self, f: F) where
F: FnOnce(&mut S, &mut dyn Context<B>) + Send + 'static,

Execute a closure with the chain-specific network specialization.

pub fn with_gossip<F>(&self, f: F) where
F: FnOnce(&mut ConsensusGossip<B>, &mut dyn Context<B>) + Send + 'static,

Execute a closure with the consensus gossip.

pub fn is_major_syncing(&self) -> bool

Are we in the process of downloading the chain?

pub fn status(&self) -> UnboundedReceiver<ProtocolStatus<B>>

Get sync status

pub fn network_state(&self) -> NetworkState

Get network state.

pub fn peers_debug_info(&self) -> Vec<(PeerId, PeerInfo<B>)>

Get currently connected peers.

Warning: This method can return outdated information and should only ever be used when obtaining outdated information is acceptable.