Module substrate_network::test

Structs

DummySpecialization
The test specialization.

EmptyTransactionPool

ForceFinalized

JustificationTestNet

NoopLink
A link implementation that does nothing.

PassThroughVerifier
A Verifier that accepts all blocks and passes them on with the configured finality to be imported.

Peer

TestLink
A Link that can wait for a block to have been imported.

TestNet

Transfer
Calls in transactions.

Enums

Extrinsic
Extrinsic for test-runtime.

Traits

SpecializationFactory

TestClient
Extension trait for a test client.

TestNetFactory

Type Definitions

Block
A test block.

Hash
A simple hash type for all our hashing.

PeersClient